Tools/PHP:
* Erase any previously created dump file before processing.
* Added a custom error handler (Kind of useless for now, just in case).
* Fixed memory usage by deallocating EAI entries once they are converted to SAI, and add various memory deallocation in places where garbage collection fails. This should allow the script to reasonably run on any machine.
* For some awkward reason, `foreach (...)` sometimes bitches. Use `for (...)` instead. This causes a performance decrease, but that's not the most important matter of this script.
* Added a flavoured progressing output line.

Note: I'm confused between `creature_text`.`id` and `creature_text`.`groupid`. @Discover- ?

+++ PHP/version2.php
@@ -65,6 +65,8 @@ foreach ($EAIDataSet as $eaiItem) {
     $npcStore[$npcId]->addEAI($eaiItem);
 }
 
+unset($eaiItem, $npcName, $npcId, $EAIDataSet); // Save some memory
+
 $storeSize = count($npcStore);
 
 ob_start();
@@ -72,12 +74,14 @@ echo '>> ' . $storeSize . ' different NPC EAIs detected in ' . round(microtime(t
 echo 'Converting ... (0%)' . PHP_EOL;
 ob_end_flush();
 
-unlink('creature_texts_v2.sql');
-unlink('smart_scripts_v2.sql');
+if (file_exists('creature_texts_v2.sql'))
+    unlink('creature_texts_v2.sql');
+if (file_exists('smart_scripts_v2.sql'))
+    unlink('smart_scripts_v2.sql');
 
 $itr = 0;
 foreach ($npcStore as $npcId => $npcObj) {
-    echo $npcObj->countSQLRows() . ' EAI sql row found for NPC ' . $npcObj->npcName . PHP_EOL;
+    // echo $npcObj->countSQLRows() . ' EAI sql row found for NPC ' . $npcObj->npcName . PHP_EOL;
 
     $npcObj->convertAllToSAI();
     $npcObj->getSmartScripts(false); // Dump texts ONLY
